requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==22.2.0
blis==0.7.10
catalogue==1.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.14
flatbuffers==23.1.4
fonttools==4.38.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
joblib==1.3.2
kiwisolver==1.4.4
langchain==0.0.271
langsmith==0.0.24
marshmallow==3.20.1
matplotlib==3.6.3
mediapipe==0.9.0.1
msvc-runtime==14.34.31931
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.4
openai==0.27.9
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
pandas==2.0.3
Pillow==9.4.0
plac==1.1.3
preshed==3.0.8
protobuf==3.20.3
pydantic==2.2.1
pydantic_core==2.6.1
PyMuPDF==1.22.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
spacy==2.3.5
SQLAlchemy==2.0.20
srsly==1.0.7
tenacity==8.2.3
textblob==0.17.1
thinc==7.4.6
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==0.10.1
yarl==1.9.2
django==4.2.4
djangorestframework